requirements.txt
Django==2.0
django_countries==5.0
Pillow==5.0.0
